requirements.txt
keras==2.7.0
numpy==1.22.0
opencv_python==4.5.5.62
Pillow==9.0.0
streamlit==1.3.1
